+++ test/casting.cpp
@@ -174,7 +174,7 @@ namespace zacc { namespace test {
         auto actual = make_mask(condition);
         auto data = actual.data();
 
-        for(int i = 0; i < size_v<TypeParam>; i++)
+        for(size_t i = 0; i < size_v<TypeParam>; i++)
         {
             VASSERT_EQ(data[i], expected[i]);
         }
